# Improve conversion performance