--- a/MANIFEST.in
+++ b/MANIFEST.in
@@ -14,5 +14,6 @@ graft include
 graft tests
 graft dev
 graft docs
+graft profiling
 graft .github
 global-exclude *.py[cod] __pycache__ *.so