--- a/include/fastnumbers/parser/character.hpp
+++ b/include/fastnumbers/parser/character.hpp
@@ -56,9 +56,8 @@ public:
         // Attempt to simply see if all the characters are digits.
         // If so it is an integer.
         const char* str = m_start;
-        while (is_valid_digit(*str)) {
-            str += 1;
-        }
+        const char* str_end = end();
+        consume_digits(str, str_end);
         return str != m_start && str == (m_start + m_str_len);
     }