… CHANGELOG.md
… README.rst
… TIMING_35.ipynb
… TIMING_36.ipynb
… TIMING_37.ipynb
… docs/source/timing.rst
… include/fastnumbers/parser/numeric.hpp
… include/fastnumbers/parser/unicode.hpp
… include/fastnumbers/third_party/fast_float.h
… profiling/README.md
… profiling/profile.py
… profiling/results-3.10.md
… profiling/results-3.7.md
… profiling/results-3.8.md
… profiling/results-3.9.md
… setup.py
… src/cpp/argparse.cpp
… src/cpp/c_str_parsing.cpp
… src/cpp/extractor.cpp
… src/cpp/fastnumbers.cpp
… src/cpp/parser.cpp
… src/fastnumbers/__init__.py
… src/fastnumbers/fastnumbers.pyi
… tests/test_fastnumbers.py
… tests/test_fastnumbers_examples.py
… tox.ini