In addition to #57, #59, this PR optimizes performance across many aspects of `fastnumbers`. Some enhancements were algorithmic, some were utilizing more efficient libraries, and some were micro-optimizations and bithacks.

This PR also solves #28 by using a more robust string to double conversion function (the one mentioned in #57).